refactor: Every type, model, schema should be in its own dedicated file/folder. With this convetion you fastly find where you need to make some code changes.  It helps in future with huge codebases.
