+++ server/models/Appointment.js
@@ -1,8 +1,10 @@
 const mongoose = require('mongoose');
+const mexp = require('mongoose-elasticsearch-xp');
+const { generate } = require('mongoose-elasticsearch-xp/lib/mapping');
 
-module.exports = new mongoose.Schema(
+const AppointmentSchema = new mongoose.Schema(
   {
-    agent_id: { type: mongoose.Schema.Types.ObjectId, ref: 'Agent'},
+    agent_id: { type: mongoose.Schema.Types.ObjectId, ref: 'Agent' },
     customer_id: { type: mongoose.Schema.Types.ObjectId, ref: 'Customer' },
     time: {
       type: Date,
@@ -13,6 +15,43 @@ module.exports = new mongoose.Schema(
   },
   {
     minimize: false,
-    timestamps: { createdAt: 'created_at', updatedAt: 'updated_at' }
+    timestamps: { createdAt: 'created_at', updatedAt: 'updated_at' },
   }
 );
+
+AppointmentSchema.plugin(mexp, {
+  client: elasticClient,
+  index: 'appointments',
+  type: 'appointment',
+});
+
+const Appointment = mongoose.model('Appointment', AppointmentSchema, 'Appointments');
+
+const AppointmentTC = composeWithMongoose(Appointment);
+
+const AppointmentEsTC = composeWithElastic({
+  graphqlTypeName: 'AppointmentES',
+  elasticIndex: 'appointments',
+  elasticType: 'appointment',
+  elasticMapping: { properties: generate(AppointmentSchema) },
+  elasticClient,
+  pluralFields: [],
+});
+
+AppointmentEsTC.getResolver('search')
+  .getTypeComposer()
+  .getFieldTC('hits')
+  .addRelation('fromMongo', {
+    resolver: () => AppointmentTC.getResolver('findById'),
+    prepareArgs: {
+      _id: source => source._id,
+    },
+    projection: { _id: true },
+  });
+
+module.exports = {
+  AppointmentSchema,
+  Appointment,
+  AppointmentTC,
+  AppointmentEsTC,
+};
